refactor: goals modules refactor
