+++ src/goals/goals.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
-import { GoalsController } from './controller/goals/goals.controller';
-import { GoalsService } from './services/goals/goals.service';
+import { GoalsController } from './controller/goals.controller';
+import { GoalsService } from './services/goals.service';
 import { SharedServicesModule } from 'src/shared/shared-services.module';
 
 @Module({
